screen-orientation/browsing-contexts.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<script>
  function loadIframe(src = "about:blank") {
    const iframe = document.createElement("iframe");
    iframe.src = src;
    document.body.appendChild(iframe);
    return new Promise(r => {
      if (iframe.contentDocument.readyState === "complete") {
        return r(iframe);
      }
      iframe.onload = () => r(iframe);
    });
  }

  promise_test(async t => {
    const iframe = await loadIframe();
    await test_driver.bless("request full screen", () => {
      return document.documentElement.requestFullscreen();
    });
    const { type: currentOrientation } = screen.orientation.type;
    const isPortrait = currentOrientation.includes("portrait");
    const newOrientation = `${isPortrait ? "landscape" : "portrait"}`;
    const promiseToLockIFrame = iframe.contentWindow.screen.orientation.lock(
      `${newOrientation}-secondary`
    );
    await screen.orientation.lock(`${newOrientation}-primary`);
    await promise_rejects(
      t,
      "AbortError",
      promiseToLockIFrame,
      "Descendant browsing context's pending promise is rejected"
    );
    screen.orientation.unlock();
    await document.exitFullscreen();
    iframe.remove();
  }, "Browsing contexts test");

  promise_test(async t => {
    const iframe = await loadIframe();
    const promiseToChange = new Promise(resolve => {
      screen.orientation.addEventListener("change", resolve);
    });
    await test_driver.bless("request full screen", () => {
      return document.documentElement.requestFullscreen();
    });
    const { type: originalOrientation } = screen.orientation.type;
    const isPortrait = originalOrientation.startsWith("portrait");
    const newOrientation = `${isPortrait ? "landscape" : "portrait"}-primary`;
    await iframe.contentWindow.screen.orientation.lock(newOrientation);
    window.location.href = "https://not-web-platform.test:8443";
    await promiseToChange;
    assert_equals(
      screen.orientation.type,
      originalOrientation,
      "Navigating from iframe to new top-level browsing context unlocks orientation"
    );
    screen.orientation.unlock();
    await document.exitFullscreen();
    iframe.remove();
  }, "Iframe navigation test");

  promise_test(async t => {
    const promiseToChange = new Promise(resolve => {
      screen.orientation.addEventListener("change", resolve);
    });
    await test_driver.bless("request full screen", () => {
      return document.documentElement.requestFullscreen();
    });
    const { type: originalOrientation } = screen.orientation.type;
    const isPortrait = originalOrientation.startsWith("portrait");
    const newOrientation = `${isPortrait ? "landscape" : "portrait"}-primary`;
    await screen.orientation.lock("newOrientation");
    window.open(
      "https://not-web-platform.test:8443",
      "popup test",
      "width=500,height=300"
    );
    await promiseToChange;
    assert_equals(
      screen.orientation.type,
      originalOrientation,
      "Opening pop-up screen unlocks orientation"
    );
    screen.orientation.unlock();
    window.close();
    await document.exitFullscreen();
  }, "Opening pop-up window test");
</script>
